refactor: diagramFilter -> infoFilter
other filters apply to the diagram outside of diagramFilter.ts,
so renaming to be more distinct.

infoFilter refers to infoCategory-based filters,
i.e. standard filters, as opposed to other diagram filters like
showImpliedEdges.
